Prototype v0.0.1
Kinda... Its a massive step in that direction. The first implementation of the Landlord Reviews page is in. Review Scores have been switched from numerical to star icons. <-- Does not yet properly reflect the score although this data is loaded from the database. With that -- Styles for the Search Results page and Landlord Reviews page have been implemented. <-- Currently only works on IOS (Andriod Not Tested) versions. For Scope reasons as we approach the prototype deadline it will remain like this.